# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MtaaReach CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete/deactivate user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── GEOGRAPHY ─────────────────────────────────────────────────────────────
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List SMS templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create SMS template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{templateId}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Archive template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /templates/{templateId}/duplicate:
    post:
      operationId: duplicateTemplate
      tags: [templates]
      summary: Duplicate a template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"

  # ── CAMPAIGNS ─────────────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, queued, sending, completed, paused, failed]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignsPage"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create and optionally queue a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign detail and stats
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign (draft only)
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a sending campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/resume:
    post:
      operationId: resumeCampaign
      tags: [campaigns]
      summary: Resume a paused campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/preview:
    post:
      operationId: previewCampaign
      tags: [campaigns]
      summary: Preview audience size and estimated cost before confirming
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignPreview"
  /sms-logs:
    get:
      operationId: listSmsLogs
      tags: [campaigns]
      summary: List SMS delivery logs (filter by campaign, status, etc.)
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [sent, delivered, failed, pending]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsLogsPage"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [super_admin, tenant_admin, county_coordinator, constituency_coordinator, ward_coordinator, village_coordinator, field_agent]
        isActive:
          type: boolean
        tenantId:
          type: ["integer", "null"]
        assignedCountyId:
          type: ["integer", "null"]
        assignedConstituencyId:
          type: ["integer", "null"]
        assignedWardId:
          type: ["integer", "null"]
        assignedVillageId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        assignedCountyId:
          type: ["integer", "null"]
        assignedConstituencyId:
          type: ["integer", "null"]
        assignedWardId:
          type: ["integer", "null"]
        assignedVillageId:
          type: ["integer", "null"]

    # GEOGRAPHY
...
    Template:
      type: object
      required: [id, name, body, tenantId, isArchived, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        body:
          type: string
        charCount:
          type: integer
        partCount:
          type: integer
        tenantId:
          type: integer
        isArchived:
          type: boolean
        createdAt:
          type: string
    TemplateInput:
      type: object
      required: [name, body]
      properties:
        name:
          type: string
        body:
          type: string
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        body:
          type: string

    # CAMPAIGNS
    Campaign:
      type: object
      required: [id, name, senderId, body, status, tenantId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        senderId:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [draft, queued, sending, completed, paused, failed]
        tenantId:
          type: integer
        templateId:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        totalRecipients:
          type: integer
        sentCount:
          type: integer
        deliveredCount:
          type: integer
        failedCount:
          type: integer
        estimatedCost:
          type: number
        actualCost:
          type: number
        audienceFilter:
          $ref: "#/components/schemas/AudienceFilter"
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name, senderId, body, audience]
      properties:
        name:
          type: string
        senderId:
          type: string
        body:
          type: string
        templateId:
          type: ["integer", "null"]
        audience:
          $ref: "#/components/schemas/AudienceFilter"
        scheduledAt:
          type: ["string", "null"]
        sendNow:
          type: boolean
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        senderId:
          type: string
        body:
          type: string
        audience:
          $ref: "#/components/schemas/AudienceFilter"
        scheduledAt:
          type: ["string", "null"]
    AudienceFilter:
      type: object
      properties:
        countyIds:
          type: array
          items:
            type: integer
        constituencyIds:
          type: array
          items:
            type: integer
        wardIds:
          type: array
          items:
            type: integer
        villageIds:
          type: array
          items:
            type: integer
        pollingStationIds:
          type: array
          items:
            type: integer
        groupIds:
          type: array
          items:
            type: integer
        contactIds:
          type: array
          items:
            type: integer
    CampaignPreview:
      type: object
      required: [totalSelected, eligible, excluded, estimatedCost, walletBalance, canAfford]
      properties:
        totalSelected:
          type: integer
        eligible:
          type: integer
        excluded:
          type: integer
        estimatedCost:
          type: number
        walletBalance:
          type: number
        canAfford:
          type: boolean
    CampaignsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    SmsLog:
      type: object
      required: [id, campaignId, contactId, phone, status, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        contactId:
          type: integer
        phone:
          type: string
        status:
          type: string
          enum: [sent, delivered, failed, pending]
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        createdAt:
          type: string
    SmsLogsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SmsLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # SENDER IDs
...