# lib/api-zod/src/generated/types/audienceFilter.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MtaaReach CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface AudienceFilter {
  countyIds?: number[];
  constituencyIds?: number[];
  wardIds?: number[];
  villageIds?: number[];
  pollingStationIds?: number[];
  groupIds?: number[];
  contactIds?: number[];
}